Valve Sequencing Addition
I changes some methods and added a few in "diagram" to test valve sequencing. One dummy sequence is implemented to test the code.

The one sequence is usable with a button. When pressed, it closes all valves and opens the ones listed in the sequence.

Moving forward, we could add each sequence in a drop-down menu or series of buttons. We could also add delays to these sequences. However, it may be necessary to add a separate component to organize all theses sequences on screen.